// src/lib/types.ts
export enum PasswordStrength {
  EASY = "EASY",
  HARD = "HARD",
  MEDIUM = "MEDIUM",
}

export enum PasswordValidations {
  DIGIT = "DIGIT",
  UPPERCASE = "UPPERCASE",
  LOWERCASE = "LOWERCASE",
  SPECIAL_CHARACTER = "SPECIAL_CHARACTER",
  AT_LEAST_8_DIGITS = "AT_LEAST_8_DIGITS",
  MORE_THAN_10_DIGITS = "MORE_THAN_10_DIGITS",
}